// lib/delhivery/types.ts
export interface DelhiveryShipment {
  name: string;
  order: string;
  phone: string;
  add: string;
  pin: number;
  payment_mode: "Prepaid" | "COD" | "Pickup" | "REPL";
  address_type?: string;
  ewbn?: string;
  hsn_code?: string;
  shipping_mode?: "Surface" | "Express";
  seller_inv?: string;
  city?: string;
  weight: number;
  return_name?: string;
  return_address?: string;
  return_city?: string;
  return_phone?: string;
  return_state?: string;
  return_country?: string;
  return_pin?: number | null;
  seller_name?: string;
  fragile_shipment?: boolean;
  shipment_height?: number;
  shipment_width?: number;
  shipment_length?: number;
  cod_amount?: number;
  products_desc?: string;
  state?: string;
  dangerous_good?: boolean;
  waybill?: string;
  total_amount?: number;
  seller_add?: string;
  country?: string;
  plastic_packaging?: boolean;
  quantity?: string;
  transport_speed?: "F" | "D";
  order_date?: string | null;
  shipment_type?: "MPS";
  mps_amount?: number;
  mps_children?: number;
  master_id?: string;
}

export interface CreateShipmentRequest {
  shipments: DelhiveryShipment[];
  pickup_location: {
    name: string;
  };
}